+++ statutory/CHANGELOG.md
@@ -1,3 +1,14 @@
+<a name="0.27.0"></a>
+# [0.27.0](https://github.com/AEGEE/oms-statutory/compare/0.26.2...0.27.0) (2019-12-22)
+
+
+### Features
+
+* **candidates:** add candidates export. Fixes MEMB-499 ([b3ed05f](https://github.com/AEGEE/oms-statutory/commit/b3ed05f))
+
+
+
+<a name="0.26.2"></a>
 ## [0.26.2](https://github.com/AEGEE/oms-statutory/compare/0.26.1...0.26.2) (2019-11-25)
 
 
@@ -7,6 +18,7 @@
 
 
 
+<a name="0.26.1"></a>
 ## [0.26.1](https://github.com/AEGEE/oms-statutory/compare/0.26.0...0.26.1) (2019-11-24)
 
 
@@ -16,6 +28,7 @@
 
 
 
+<a name="0.26.0"></a>
 # [0.26.0](https://github.com/AEGEE/oms-statutory/compare/0.25.3...0.26.0) (2019-11-24)
 
 
@@ -30,6 +43,7 @@
 
 
 
+<a name="0.25.3"></a>
 ## [0.25.3](https://github.com/AEGEE/oms-statutory/compare/0.25.2...0.25.3) (2019-11-24)
 
 
@@ -39,6 +53,7 @@
 
 
 
+<a name="0.25.2"></a>
 ## [0.25.2](https://github.com/AEGEE/oms-statutory/compare/0.25.1...0.25.2) (2019-11-13)
 
 
@@ -48,6 +63,7 @@
 
 
 
+<a name="0.25.1"></a>
 ## [0.25.1](https://github.com/AEGEE/oms-statutory/compare/0.25.0...0.25.1) (2019-11-11)
 
 
@@ -58,6 +74,7 @@
 
 
 
+<a name="0.25.0"></a>
 # [0.25.0](https://github.com/AEGEE/oms-statutory/compare/0.24.2...0.25.0) (2019-11-09)
 
 
@@ -67,6 +84,7 @@
 
 
 
+<a name="0.24.2"></a>
 ## [0.24.2](https://github.com/AEGEE/oms-statutory/compare/0.24.1...0.24.2) (2019-11-06)
 
 
@@ -76,6 +94,7 @@
 
 
 
+<a name="0.24.1"></a>
 ## [0.24.1](https://github.com/AEGEE/oms-statutory/compare/0.23.1...0.24.1) (2019-11-05)
 
 
@@ -93,14 +112,24 @@
 
 
 
-## [0.23.1](https://github.com/AEGEE/oms-statutory/compare/0.22.6...0.23.1) (2019-11-04)
+<a name="0.23.1"></a>
+## [0.23.1](https://github.com/AEGEE/oms-statutory/compare/0.23.0...0.23.1) (2019-11-04)
+
+
+### Bug Fixes
+
+* **general:** added host for config ([e9a3d74](https://github.com/AEGEE/oms-statutory/commit/e9a3d74))
+
+
+
+<a name="0.23.0"></a>
+# [0.23.0](https://github.com/AEGEE/oms-statutory/compare/0.22.6...0.23.0) (2019-11-04)
 
 
 ### Bug Fixes
 
 * **bugsnag:** add bugsnag params. Fixes MEMB-642 ([788000a](https://github.com/AEGEE/oms-statutory/commit/788000a))
 * **docker:** renamed dockerfile ([3368a86](https://github.com/AEGEE/oms-statutory/commit/3368a86))
-* **general:** added host for config ([e9a3d74](https://github.com/AEGEE/oms-statutory/commit/e9a3d74))
 * **general:** removed config example. Fixes MEMB-687 ([81bde18](https://github.com/AEGEE/oms-statutory/commit/81bde18))
 
 
@@ -110,24 +139,18 @@
 
 
 
-## [0.22.6](https://github.com/AEGEE/oms-statutory/compare/0.22.5...0.22.6) (2019-10-25)
+<a name="0.22.6"></a>
+## [0.22.6](https://github.com/AEGEE/oms-statutory/compare/0.22.4...0.22.6) (2019-10-25)
 
 
 ### Bug Fixes
 
 * **question:** fixed questions sorting again ([052690b](https://github.com/AEGEE/oms-statutory/commit/052690b))
-
-
-
-## [0.22.5](https://github.com/AEGEE/oms-statutory/compare/0.22.4...0.22.5) (2019-10-25)
-
-
-### Bug Fixes
-
 * **questions:** fixed questions sorting on question lines listing ([37fdc9e](https://github.com/AEGEE/oms-statutory/commit/37fdc9e))
 
 
 
+<a name="0.22.4"></a>
 ## [0.22.4](https://github.com/AEGEE/oms-statutory/compare/0.22.3...0.22.4) (2019-10-13)
 
 
@@ -137,6 +160,7 @@
 
 
 
+<a name="0.22.3"></a>
 ## [0.22.3](https://github.com/AEGEE/oms-statutory/compare/0.22.2...0.22.3) (2019-10-13)
 
 
@@ -148,6 +172,7 @@
 
 
 
+<a name="0.22.2"></a>
 ## [0.22.2](https://github.com/AEGEE/oms-statutory/compare/0.22.1...0.22.2) (2019-09-26)
 
 
@@ -158,6 +183,7 @@
 
 
 
+<a name="0.22.1"></a>
 ## [0.22.1](https://github.com/AEGEE/oms-statutory/compare/0.22.0...0.22.1) (2019-09-24)
 
 
@@ -167,6 +193,7 @@
 
 
 
+<a name="0.22.0"></a>
 # [0.22.0](https://github.com/AEGEE/oms-statutory/compare/0.21.5...0.22.0) (2019-09-22)
 
 
@@ -176,6 +203,7 @@
 
 
 
+<a name="0.21.5"></a>
 ## [0.21.5](https://github.com/AEGEE/oms-statutory/compare/0.21.4...0.21.5) (2019-09-21)
 
 
@@ -185,6 +213,7 @@
 
 
 
+<a name="0.21.4"></a>
 ## [0.21.4](https://github.com/AEGEE/oms-statutory/compare/0.21.3...0.21.4) (2019-09-19)
 
 
@@ -194,6 +223,7 @@
 
 
 
+<a name="0.21.3"></a>
 ## [0.21.3](https://github.com/AEGEE/oms-statutory/compare/0.21.2...0.21.3) (2019-09-18)
 
 
@@ -203,6 +233,7 @@
 
 
 
+<a name="0.21.2"></a>
 ## [0.21.2](https://github.com/AEGEE/oms-statutory/compare/0.21.1...0.21.2) (2019-09-18)
 
 
@@ -212,6 +243,7 @@
 
 
 
+<a name="0.21.1"></a>
 ## [0.21.1](https://github.com/AEGEE/oms-statutory/compare/0.21.0...0.21.1) (2019-09-18)
 
 
@@ -222,6 +254,7 @@
 
 
 
+<a name="0.21.0"></a>
 # [0.21.0](https://github.com/AEGEE/oms-statutory/compare/0.20.2...0.21.0) (2019-09-16)
 
 
@@ -231,6 +264,7 @@
 
 
 
+<a name="0.20.2"></a>
 ## [0.20.2](https://github.com/AEGEE/oms-statutory/compare/0.20.1...0.20.2) (2019-09-11)
 
 
@@ -240,6 +274,7 @@
 
 
 
+<a name="0.20.1"></a>
 ## [0.20.1](https://github.com/AEGEE/oms-statutory/compare/0.20.0...0.20.1) (2019-09-11)
 
 
@@ -250,6 +285,7 @@
 
 
 
+<a name="0.20.0"></a>
 # [0.20.0](https://github.com/AEGEE/oms-statutory/compare/0.19.0...0.20.0) (2019-09-07)
 
 
@@ -265,6 +301,7 @@
 
 
 
+<a name="0.19.0"></a>
 # [0.19.0](https://github.com/AEGEE/oms-statutory/compare/0.18.0...0.19.0) (2019-08-28)
 
 
@@ -274,6 +311,7 @@
 
 
 
+<a name="0.18.0"></a>
 # [0.18.0](https://github.com/AEGEE/oms-statutory/compare/0.17.0...0.18.0) (2019-08-27)
 
 
@@ -283,6 +321,7 @@
 
 
 
+<a name="0.17.0"></a>
 # [0.17.0](https://github.com/AEGEE/oms-statutory/compare/0.16.4...0.17.0) (2019-08-26)
 
 
@@ -297,6 +336,7 @@
 
 
 
+<a name="0.16.4"></a>
 ## [0.16.4](https://github.com/AEGEE/oms-statutory/compare/0.16.3...0.16.4) (2019-08-25)
 
 
@@ -306,6 +346,7 @@
 
 
 
+<a name="0.16.3"></a>
 ## [0.16.3](https://github.com/AEGEE/oms-statutory/compare/0.16.2...0.16.3) (2019-08-25)
 
 
@@ -315,6 +356,7 @@
 
 
 
+<a name="0.16.2"></a>
 ## [0.16.2](https://github.com/AEGEE/oms-statutory/compare/0.16.1...0.16.2) (2019-08-24)
 
 
@@ -325,6 +367,7 @@
 
 
 
+<a name="0.16.1"></a>
 ## [0.16.1](https://github.com/AEGEE/oms-statutory/compare/0.16.0...0.16.1) (2019-08-23)
 
 
@@ -334,34 +377,24 @@
 
 
 
-# [0.16.0](https://github.com/AEGEE/oms-statutory/compare/0.15.0...0.16.0) (2019-08-21)
-
-
-### Bug Fixes
-
-* **tests:** fix another flapping test ([359bb21](https://github.com/AEGEE/oms-statutory/commit/359bb21))
-
-
-### Features
-
-* **statutory:** added /tasks endpoint. Fixes MEMB-594 ([460e00f](https://github.com/AEGEE/oms-statutory/commit/460e00f))
-
-
-
-# [0.15.0](https://github.com/AEGEE/oms-statutory/compare/0.14.2...0.15.0) (2019-08-19)
+<a name="0.16.0"></a>
+# [0.16.0](https://github.com/AEGEE/oms-statutory/compare/0.14.2...0.16.0) (2019-08-21)
 
 
 ### Bug Fixes
 
 * **deps:** updated Jest to the latest ([8ef6f55](https://github.com/AEGEE/oms-statutory/commit/8ef6f55))
+* **tests:** fix another flapping test ([359bb21](https://github.com/AEGEE/oms-statutory/commit/359bb21))
 
 
 ### Features
 
 * **question lines:** added question lines management. Fixes MEMB-590 ([7d11ef1](https://github.com/AEGEE/oms-statutory/commit/7d11ef1))
+* **statutory:** added /tasks endpoint. Fixes MEMB-594 ([460e00f](https://github.com/AEGEE/oms-statutory/commit/460e00f))
 
 
 
+<a name="0.14.2"></a>
 ## [0.14.2](https://github.com/AEGEE/oms-statutory/compare/0.14.1...0.14.2) (2019-08-17)
 
 
@@ -375,6 +408,7 @@
 
 
 
+<a name="0.14.1"></a>
 ## [0.14.1](https://github.com/AEGEE/oms-statutory/compare/0.14.0...0.14.1) (2019-08-10)
 
 
@@ -389,6 +423,7 @@
 
 
 
+<a name="0.14.0"></a>
 # [0.14.0](https://github.com/AEGEE/oms-statutory/compare/0.13.0...0.14.0) (2019-07-30)
 
 
@@ -412,6 +447,7 @@
 
 
 
+<a name="0.13.0"></a>
 # [0.13.0](https://github.com/AEGEE/oms-statutory/compare/0.12.1...0.13.0) (2019-06-02)
 
 
@@ -427,6 +463,7 @@
 
 
 
+<a name="0.12.1"></a>
 ## [0.12.1](https://github.com/AEGEE/oms-statutory/compare/0.12.0...0.12.1) (2019-05-26)
 
 
@@ -436,6 +473,7 @@
 
 
 
+<a name="0.12.0"></a>
 # [0.12.0](https://github.com/AEGEE/oms-statutory/compare/0.11.1...0.12.0) (2019-05-25)
 
 
@@ -451,6 +489,7 @@
 
 
 
+<a name="0.11.1"></a>
 ## [0.11.1](https://github.com/AEGEE/oms-statutory/compare/0.11.0...0.11.1) (2019-05-23)
 
 
@@ -460,6 +499,7 @@
 
 
 
+<a name="0.11.0"></a>
 # [0.11.0](https://github.com/AEGEE/oms-statutory/compare/0.10.8...0.11.0) (2019-05-21)
 
 
@@ -470,6 +510,7 @@
 
 
 
+<a name="0.10.8"></a>
 ## [0.10.8](https://github.com/AEGEE/oms-statutory/compare/0.10.7...0.10.8) (2019-05-19)
 
 
@@ -479,6 +520,7 @@
 
 
 
+<a name="0.10.7"></a>
 ## [0.10.7](https://github.com/AEGEE/oms-statutory/compare/0.10.6...0.10.7) (2019-05-19)
 
 
@@ -488,6 +530,7 @@
 
 
 
+<a name="0.10.6"></a>
 ## [0.10.6](https://github.com/AEGEE/oms-statutory/compare/0.10.5...0.10.6) (2019-05-19)
 
 
@@ -497,6 +540,7 @@
 
 
 
+<a name="0.10.5"></a>
 ## [0.10.5](https://github.com/AEGEE/oms-statutory/compare/0.10.4...0.10.5) (2019-04-29)
 
 
@@ -506,6 +550,7 @@
 
 
 
+<a name="0.10.4"></a>
 ## [0.10.4](https://github.com/AEGEE/oms-statutory/compare/0.10.3...0.10.4) (2019-04-29)
 
 
@@ -515,6 +560,7 @@
 
 
 
+<a name="0.10.3"></a>
 ## [0.10.3](https://github.com/AEGEE/oms-statutory/compare/0.10.2...0.10.3) (2019-04-29)
 
 
@@ -524,6 +570,7 @@
 
 
 
+<a name="0.10.2"></a>
 ## [0.10.2](https://github.com/AEGEE/oms-statutory/compare/0.10.1...0.10.2) (2019-04-28)
 
 
@@ -534,6 +581,7 @@
 
 
 
+<a name="0.10.1"></a>
 ## [0.10.1](https://github.com/AEGEE/oms-statutory/compare/0.10.0...0.10.1) (2019-04-28)
 
 
@@ -543,6 +591,7 @@
 
 
 
+<a name="0.10.0"></a>
 # [0.10.0](https://github.com/AEGEE/oms-statutory/compare/0.9.12...0.10.0) (2019-04-28)
 
 
@@ -552,6 +601,7 @@
 
 
 
+<a name="0.9.12"></a>
 ## [0.9.12](https://github.com/AEGEE/oms-statutory/compare/0.9.11...0.9.12) (2019-04-28)
 
 
@@ -561,6 +611,7 @@
 
 
 
+<a name="0.9.11"></a>
 ## [0.9.11](https://github.com/AEGEE/oms-statutory/compare/0.9.10...0.9.11) (2019-04-23)
 
 
@@ -570,6 +621,7 @@
 
 
 
+<a name="0.9.10"></a>
 ## [0.9.10](https://github.com/AEGEE/oms-statutory/compare/0.9.9...0.9.10) (2019-04-18)
 
 
@@ -579,6 +631,7 @@
 
 
 
+<a name="0.9.9"></a>
 ## [0.9.9](https://github.com/AEGEE/oms-statutory/compare/0.9.8...0.9.9) (2019-04-16)
 
 
@@ -589,6 +642,7 @@
 
 
 
+<a name="0.9.8"></a>
 ## [0.9.8](https://github.com/AEGEE/oms-statutory/compare/0.9.7...0.9.8) (2019-04-16)
 
 
@@ -600,6 +654,7 @@
 
 
 
+<a name="0.9.7"></a>
 ## [0.9.7](https://github.com/AEGEE/oms-statutory/compare/0.9.6...0.9.7) (2019-04-14)
 
 
@@ -615,6 +670,7 @@
 
 
 
+<a name="0.9.6"></a>
 ## [0.9.6](https://github.com/AEGEE/oms-statutory/compare/0.9.5...0.9.6) (2019-04-10)
 
 
@@ -624,6 +680,7 @@
 
 
 
+<a name="0.9.5"></a>
 ## [0.9.5](https://github.com/AEGEE/oms-statutory/compare/0.9.4...0.9.5) (2019-04-09)
 
 
@@ -634,6 +691,7 @@
 
 
 
+<a name="0.9.4"></a>
 ## [0.9.4](https://github.com/AEGEE/oms-statutory/compare/0.9.3...0.9.4) (2019-04-09)
 
 
@@ -648,6 +706,7 @@
 
 
 
+<a name="0.9.3"></a>
 ## [0.9.3](https://github.com/AEGEE/oms-statutory/compare/0.9.2...0.9.3) (2019-04-08)
 
 
@@ -657,6 +716,7 @@
 
 
 
+<a name="0.9.2"></a>
 ## [0.9.2](https://github.com/AEGEE/oms-statutory/compare/0.9.1...0.9.2) (2019-04-08)
 
 
@@ -666,6 +726,7 @@
 
 
 
+<a name="0.9.1"></a>
 ## [0.9.1](https://github.com/AEGEE/oms-statutory/compare/0.9.0...0.9.1) (2019-03-31)
 
 
@@ -675,6 +736,7 @@
 
 
 
+<a name="0.9.0"></a>
 # [0.9.0](https://github.com/AEGEE/oms-statutory/compare/0.8.0...0.9.0) (2019-03-30)
 
 
@@ -689,6 +751,7 @@
 
 
 
+<a name="0.8.0"></a>
 # [0.8.0](https://github.com/AEGEE/oms-statutory/compare/be680b8...0.8.0) (2019-03-24)